Updates to Projectile Path
@enitaity Please add a Projectile object to Weapon and a getProjectile() method. Also, the math does not work... Remember that the mouse position being fed in is absolute! Finally, your getFullPath() method throws ArraysOutOfBoundException.

+++ Main/src/Weapons.java
@@ -83,13 +83,13 @@ public class Weapons {
     return ret;
   }
 
-  public double[][] getPathFull() {
-    double[][] ret = new double[3][20];
+  public int[][] getPathFull() {
+    int[][] ret = new int[3][20];
     for (int i = 1; i < 21; i++) {
       double temp = (range / 20) * i;
-      ret[0][i - 1] = xPos + temp;
-      ret[1][i - 1] = yPos + temp * Math.tan(theta) - (G * temp * temp / (2 * v0 * v0 * Math.cos(theta) * Math.cos(theta)));
-      ret[2][i - 3] = 180 * Math.atan(Math.tan(theta) - (2 * temp * G) / (2 * v0 * v0 * Math.pow(Math.cos(theta), 2)));
+      ret[0][i - 1] = (int) (xPos + temp);
+      ret[1][i - 1] = (int) (yPos + temp * Math.tan(theta) - (G * temp * temp / (2 * v0 * v0 * Math.cos(theta) * Math.cos(theta))));
+      ret[2][i - 3] = (int) (180 * Math.atan(Math.tan(theta) - (2 * temp * G) / (2 * v0 * v0 * Math.pow(Math.cos(theta), 2))));
     }
     return ret;
   }
